pyproject.toml
[tool.poetry]
name = "stamm-drift-detectors"
version = "0.1.1"
description = "STAMM Drift Detectors"
authors = [
    "Matthew Crowther <nmc215ncl.ac.uk>",
    "Brett Metcalfe <brett.metcalfe@hotmail.co.uk>",
    "Jasper Koehorst <jasper.koehorst@wur.nl>",
    "David Camilo Corrales <David-Camilo.Corrales-Munoz@inrae.fr>",
    "Carlos Alberto Suarez Muñoz <csuarez@gmail.com>"
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://gitlab.com/stamm-4m/drift_detectors_pack"

packages = [
    { include = "drift_detectors" }
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "2.2.5"
PyYAML = "6.0.2"
river = "0.22.0"
scikit_learn = "1.6.1"
scipy = "1.15.2"
setuptools = "75.1.0"

[tool.poetry.group.test.dependencies]
pytest-xdist = "^3.6.1"
pytest = "^8.3.4"
coverage = "^7.6.8"
mypy = "^1.14.1"

[tool.poetry.group.dev.dependencies]
bump2version = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
